requirements.txt
# requirements.txt
python-dotenv
streamlit
langchain
langchain-community
langchain-core
requests
pandas
Pillow
langchain-gemini
faiss-cpu
langchain-google-genai
streamlit
google.generativeai
mcp[cli]>=1.13.1
fastmcp>=0.1.0
jsonpatch
pydantic

